webhooks-with-cloud-run/MicroServices/GitHub/requirements.txt
Flask==1.1.1
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn==19.9.0
google-cloud-secret-manager==0.1.0
google-cloud-bigquery==1.23.1
jwt==0.6.1